rename `src`->`internal`
