+++ internal/app/connection/eventstream/sse_read.go
